+++ cdap-standalone/src/main/java/co/cask/cdap/StandaloneMain.java
@@ -1,5 +1,5 @@
 /*
- * Copyright © 2014-2015 Cask Data, Inc.
+ * Copyright © 2014-2016 Cask Data, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy of
@@ -21,6 +21,7 @@ import co.cask.cdap.app.guice.AppFabricServiceRuntimeModule;
 import co.cask.cdap.app.guice.ProgramRunnerRuntimeModule;
 import co.cask.cdap.app.guice.ServiceStoreModules;
 import co.cask.cdap.app.store.ServiceStore;
+import co.cask.cdap.common.ServiceBindException;
 import co.cask.cdap.common.conf.CConfiguration;
 import co.cask.cdap.common.conf.Constants;
 import co.cask.cdap.common.guice.ConfigModule;
@@ -28,6 +29,8 @@ import co.cask.cdap.common.guice.DiscoveryRuntimeModule;
 import co.cask.cdap.common.guice.IOModule;
 import co.cask.cdap.common.guice.LocationRuntimeModule;
 import co.cask.cdap.common.io.URLConnections;
+import co.cask.cdap.common.namespace.guice.NamespaceClientRuntimeModule;
+import co.cask.cdap.common.startup.ConfigurationLogger;
 import co.cask.cdap.common.utils.DirUtils;
 import co.cask.cdap.common.utils.OSDetector;
 import co.cask.cdap.data.runtime.DataFabricModules;
@@ -36,6 +39,7 @@ import co.cask.cdap.data.runtime.DataSetsModules;
 import co.cask.cdap.data.stream.StreamAdminModules;
 import co.cask.cdap.data.stream.service.StreamService;
 import co.cask.cdap.data.stream.service.StreamServiceRuntimeModule;
+import co.cask.cdap.data.view.ViewAdminModules;
 import co.cask.cdap.data2.datafabric.dataset.service.DatasetService;
 import co.cask.cdap.explore.client.ExploreClient;
 import co.cask.cdap.explore.executor.ExploreExecutorService;
@@ -47,6 +51,8 @@ import co.cask.cdap.gateway.router.RouterModules;
 import co.cask.cdap.internal.app.services.AppFabricServer;
 import co.cask.cdap.logging.appender.LogAppenderInitializer;
 import co.cask.cdap.logging.guice.LoggingModules;
+import co.cask.cdap.metadata.MetadataService;
+import co.cask.cdap.metadata.MetadataServiceModule;
 import co.cask.cdap.metrics.guice.MetricsClientRuntimeModule;
 import co.cask.cdap.metrics.guice.MetricsHandlerModule;
 import co.cask.cdap.metrics.query.MetricsQueryService;
@@ -54,8 +60,10 @@ import co.cask.cdap.notifications.feeds.guice.NotificationFeedServiceRuntimeModu
 import co.cask.cdap.notifications.guice.NotificationServiceRuntimeModule;
 import co.cask.cdap.security.guice.SecurityModules;
 import co.cask.cdap.security.server.ExternalAuthenticationServer;
+import co.cask.cdap.store.guice.NamespaceStoreModule;
 import co.cask.tephra.inmemory.InMemoryTransactionService;
 import com.google.common.base.Joiner;
+import com.google.common.base.Throwables;
 import com.google.common.collect.ImmutableList;
 import com.google.common.util.concurrent.Service;
 import com.google.inject.Guice;
@@ -88,9 +96,9 @@ public class StandaloneMain {
   private final ServiceStore serviceStore;
   private final StreamService streamService;
   private final MetricsCollectionService metricsCollectionService;
-
   private final LogAppenderInitializer logAppenderInitializer;
   private final InMemoryTransactionService txService;
+  private final MetadataService metadataService;
   private final boolean securityEnabled;
   private final boolean sslEnabled;
   private final CConfiguration configuration;
@@ -135,6 +143,7 @@ public class StandaloneMain {
     }
 
     exploreClient = injector.getInstance(ExploreClient.class);
+    metadataService = injector.getInstance(MetadataService.class);
 
     Runtime.getRuntime().addShutdownHook(new Thread() {
       @Override
@@ -160,6 +169,8 @@ public class StandaloneMain {
 
     cleanupTempDir();
 
+    ConfigurationLogger.logImportantConfig(configuration);
+
     // Start all the services.
     txService.startAndWait();
     metricsCollectionService.startAndWait();
@@ -178,6 +189,7 @@ public class StandaloneMain {
 
     metricsQueryService.startAndWait();
     router.startAndWait();
+
     if (userInterfaceService != null) {
       userInterfaceService.startAndWait();
     }
@@ -189,6 +201,7 @@ public class StandaloneMain {
     if (exploreExecutorService != null) {
       exploreExecutorService.startAndWait();
     }
+    metadataService.startAndWait();
 
     String protocol = sslEnabled ? "https" : "http";
     int dashboardPort = sslEnabled ?
@@ -230,6 +243,7 @@ public class StandaloneMain {
         externalAuthenticationServer.stopAndWait();
       }
       logAppenderInitializer.close();
+      metadataService.stopAndWait();
 
     } catch (Throwable e) {
       LOG.error("Exception during shutdown", e);
@@ -244,6 +258,11 @@ public class StandaloneMain {
   private void cleanupTempDir() {
     File tmpDir = new File(configuration.get(Constants.CFG_LOCAL_DATA_DIR),
                            configuration.get(Constants.AppFabric.TEMP_DIR)).getAbsoluteFile();
+
+    if (!tmpDir.isDirectory()) {
+      return;
+    }
+
     try {
       DirUtils.deleteDirectoryContents(tmpDir, true);
     } catch (IOException e) {
@@ -264,8 +283,17 @@ public class StandaloneMain {
       }
       main.startUp();
     } catch (Throwable e) {
-      System.err.println("Failed to start Standalone CDAP. " + e.getMessage());
-      LOG.error("Failed to start Standalone CDAP", e);
+      Throwable rootCause = Throwables.getRootCause(e);
+      if (rootCause instanceof ServiceBindException) {
+        LOG.error("Failed to start Standalone CDAP: {}", rootCause.getMessage());
+        System.err.println("Failed to start Standalone CDAP: " + rootCause.getMessage());
+      } else {
+        // exception stack trace will be logged by
+        // UncaughtExceptionIdleService.UNCAUGHT_EXCEPTION_HANDLER
+        LOG.error("Failed to start Standalone CDAP");
+        System.err.println("Failed to start Standalone CDAP");
+        e.printStackTrace(System.err);
+      }
       Runtime.getRuntime().halt(-2);
     }
   }
@@ -315,13 +343,21 @@ public class StandaloneMain {
   private static List<Module> createPersistentModules(CConfiguration configuration, Configuration hConf) {
     configuration.setIfUnset(Constants.CFG_DATA_LEVELDB_DIR, Constants.DEFAULT_DATA_LEVELDB_DIR);
 
-    String environment =
-      configuration.get(Constants.CFG_APPFABRIC_ENVIRONMENT, Constants.DEFAULT_APPFABRIC_ENVIRONMENT);
-    if (environment.equals("vpc")) {
-      System.err.println("Application Server Environment: " + environment);
-    }
-
     configuration.set(Constants.CFG_DATA_INMEMORY_PERSISTENCE, Constants.InMemoryPersistenceType.LEVELDB.name());
+
+    // configure all services except for router to bind to 127.0.0.1
+    configuration.set(Constants.AppFabric.SERVER_ADDRESS, "127.0.0.1");
+    configuration.set(Constants.Transaction.Container.ADDRESS, "127.0.0.1");
+    configuration.set(Constants.Dataset.Manager.ADDRESS, "127.0.0.1");
+    configuration.set(Constants.Dataset.Executor.ADDRESS, "127.0.0.1");
+    configuration.set(Constants.Stream.ADDRESS, "127.0.0.1");
+    configuration.set(Constants.Metrics.ADDRESS, "127.0.0.1");
+    configuration.set(Constants.Metrics.SERVER_ADDRESS, "127.0.0.1");
+    configuration.set(Constants.MetricsProcessor.ADDRESS, "127.0.0.1");
+    configuration.set(Constants.LogSaver.ADDRESS, "127.0.0.1");
+    configuration.set(Constants.Security.AUTH_SERVER_BIND_ADDRESS, "127.0.0.1");
+    configuration.set(Constants.Explore.SERVER_ADDRESS, "127.0.0.1");
+    configuration.set(Constants.Metadata.SERVICE_BIND_ADDRESS, "127.0.0.1");
 
     return ImmutableList.of(
       new ConfigModule(configuration, hConf),
@@ -344,7 +380,11 @@ public class StandaloneMain {
       new ExploreClientModule(),
       new NotificationFeedServiceRuntimeModule().getStandaloneModules(),
       new NotificationServiceRuntimeModule().getStandaloneModules(),
-      new StreamAdminModules().getStandaloneModules()
+      new ViewAdminModules().getStandaloneModules(),
+      new StreamAdminModules().getStandaloneModules(),
+      new NamespaceClientRuntimeModule().getStandaloneModules(),
+      new NamespaceStoreModule().getStandaloneModules(),
+      new MetadataServiceModule()
     );
   }
 }
